moved Reply and Response to components folder and fixed any broken links that relied upon them.
